webapp/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
PyYAML==6.0.1
